add @license with URL to metadata headers
Add @license field with license name and URL to text of the license to
metadata headers of user scripts.  Add the URL to @license in styles.
Format is kinda like what's recommended in

  https://sourceforge.net/p/greasemonkey/wiki/Metadata_Block/#license

but MIT license does not have an official homepage itself.  So just link
to the actual text of the license used, i.e. the LICENSE.txt file in the
repository on GitHub.