reverse-event-lists: don't run on pages without jQuery
Right now the script donations-reverse-event-lists.user.js depends on
jQuery version loaded by the website itself.  Some pages don't load the
library and this user scripts fails with errors.

So instead of just reporting that jQuery wasn't found, stop execution of
the user script if this isn't one of the pages that script supports.
Note that if the TODO comment is to be implemented, the script will have
to load its own jQuery via @require header, like so:

    // @require      http://code.jquery.com/jquery-3.6.0.min.js