*: add support for tracker.bsgmarathon.com
BSG Marathon has moved their tracker from bsgmarathon.com/tracker to
a subdomain tracker.bsgmarathon.com.  Add corresponding `@match` and
`url-prefix` rules to scripts and styles, where it's relevant.

A special case is donations-index-paging.user.js, which is now needed on
the new BSG website, but wasn't needed on the old version.

An exception is donations-auto-refresh-process-bids.user.js, which
doesn't support BSG website, because there is a separate
donations-bsg-autorefresh.user.js for the BSG website.  This exception
was actual on the old tracker, but might not be anymore

I don't have access to the admin parts of the tracker, so I can't check
most of the affected functionality, but I expect HTML layout and other
things to be similar enough for scripts and styles to work.